requirements.txt
aiogram
telethon
requests
python-dotenv
